requirements.txt
mysql-replication==0.18
pyyaml==3.13
redis==2.10.6
